+++ app/jobs/periodic/auth-tokens.ts
@@ -1,6 +1,7 @@
 import config from 'config';
 
-import { dbAdapter } from '../../models';
+import { dbAdapter, JobManager } from '../../models';
+import FreefeedApp from '../../freefeed-app';
 
 import { definePeriodicJob } from '.';
 
@@ -8,12 +9,13 @@ export const PERIODIC_INACTIVATE_APP_TOKENS = 'PERIODIC_INACTIVATE_APP_TOKENS';
 export const PERIODIC_REAUTH_REALTIME = 'PERIODIC_REAUTH_REALTIME';
 export const PERIODIC_CLEAN_AUTH_SESSIONS = 'PERIODIC_CLEAN_AUTH_SESSIONS';
 
-export function initHandlers(jobManager, app) {
+export function initHandlers(jobManager: JobManager, app: FreefeedApp) {
   return Promise.all([
     definePeriodicJob(jobManager, {
       name: PERIODIC_INACTIVATE_APP_TOKENS,
       handler: () => dbAdapter.periodicInvalidateAppTokens(),
-      nextTime: () => new Date(Date.now() + 10 * 60 * 1000), // every 10 minutes
+      nextTime: 10 * 60, // every 10 minutes
+      payload: {},
     }),
     definePeriodicJob(jobManager, {
       name: PERIODIC_CLEAN_AUTH_SESSIONS,
@@ -22,12 +24,14 @@ export function initHandlers(jobManager, app) {
           config.authSessions.activeSessionTTLDays,
           config.authSessions.inactiveSessionTTLDays,
         ),
-      nextTime: () => new Date(Date.now() + config.authSessions.cleanupIntervalSec * 1000),
+      nextTime: config.authSessions.cleanupIntervalSec,
+      payload: {},
     }),
     definePeriodicJob(jobManager, {
       name: PERIODIC_REAUTH_REALTIME,
       handler: () => app.context.pubsub.reAuthorizeSockets(),
-      nextTime: () => new Date(Date.now() + 5 * 60 * 1000), // every 5 minutes
+      nextTime: 5 * 60, // every 5 minutes
+      payload: {},
     }),
   ]);
 }